+++ colour/characterisation/datasets/__init__.py
@@ -2,13 +2,14 @@
 
 from __future__ import absolute_import
 
+from .aces_it import ACES_RICD
 from .cameras import CAMERA_RGB_SPECTRAL_SENSITIVITIES
 from .colour_checkers import COLOURCHECKERS, ColourChecker, COLOURCHECKER_SDS
 from .displays import DISPLAY_RGB_PRIMARIES
 from .filters import FILTER_SDS
 from .lenses import LENS_SDS
 
-__all__ = []
+__all__ = ['ACES_RICD']
 __all__ += ['CAMERA_RGB_SPECTRAL_SENSITIVITIES']
 __all__ += ['COLOURCHECKERS', 'ColourChecker', 'COLOURCHECKER_SDS']
 __all__ += ['DISPLAY_RGB_PRIMARIES']
