+++ colour/characterisation/datasets/aces_it.py
@@ -3,7 +3,7 @@
 Academy Color Encoding System - Input Transform Dataset
 =======================================================
 
-Defines the *Academy Color Encoding System* (ACES) *Input Transform* datasets.
+Defines the *Academy Color Encoding System* (ACES) *Input Transform* dataset.
 
 References
 ----------
