+++ colour/models/__init__.py
@@ -247,6 +247,10 @@ API_CHANGES : dict
 # v0.3.16
 API_CHANGES['ObjectRenamed'] = API_CHANGES['ObjectRenamed'] + [
     [
+        'colour.models.ACES_RICD',
+        'colour.characterisation.ACES_RICD',
+    ],
+    [
         'colour.models.oetf_BT2020',
         'colour.models.eotf_inverse_BT2020',
     ],
