+++ colour/models/rgb/__init__.py
@@ -13,7 +13,6 @@ from . import transfer_functions
 from .datasets import *  # noqa
 from . import datasets
 from .common import XYZ_to_sRGB, sRGB_to_XYZ
-from .aces_it import sd_to_aces_relative_exposure_values
 from .cylindrical import RGB_to_HSV, HSV_to_RGB, RGB_to_HSL, HSL_to_RGB
 from .cmyk import RGB_to_CMY, CMY_to_RGB, CMY_to_CMYK, CMYK_to_CMY
 from .prismatic import RGB_to_Prismatic, Prismatic_to_RGB
@@ -32,7 +31,6 @@ __all__ += ['RGB_to_RGB_matrix', 'RGB_to_RGB']
 __all__ += transfer_functions.__all__
 __all__ += datasets.__all__
 __all__ += ['XYZ_to_sRGB', 'sRGB_to_XYZ']
-__all__ += ['sd_to_aces_relative_exposure_values']
 __all__ += ['RGB_to_HSV', 'HSV_to_RGB', 'RGB_to_HSL', 'HSL_to_RGB']
 __all__ += ['RGB_to_CMY', 'CMY_to_RGB', 'CMY_to_CMYK', 'CMYK_to_CMY']
 __all__ += ['RGB_to_Prismatic', 'Prismatic_to_RGB']
