+++ docs/colour.characterisation.rst
@@ -3,6 +3,29 @@ Colour Characterisation
 
 .. contents:: :local:
 
+ACES Spectral Conversion
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+``colour``
+
+.. currentmodule:: colour
+
+.. autosummary::
+    :toctree: generated/
+
+    sd_to_aces_relative_exposure_values
+
+**Ancillary Objects**
+
+``colour.characterisation``
+
+.. currentmodule:: colour.characterisation
+
+.. autosummary::
+    :toctree: generated/
+
+    ACES_RICD
+
 Colour Fitting
 --------------
